+++ Assets/Assets/Scripts/PlayerManager.cs
@@ -116,7 +116,7 @@ public abstract class PlayerManager : MonoBehaviour
 
     public abstract void Finish_DealDamage_Attack1();
 
-    public abstract void OnStateReceived(float timestamp, Vector3 position, Vector3 direction, Vector3 deltaPos);
+    public abstract void OnStateReceived(long timestamp, Vector3 position, Vector3 direction, Vector3 deltaPos);
 
     public Vector3 ClampPosition(Vector3 position)
     {
